Revert "simplify tictactoe.py"
This reverts commit 69f30f5ee1a2d6a0a235961edc51da30171949fe.
