+++ charbot/tictactoe.py
@@ -23,13 +23,17 @@
 # SOFTWARE.
 #  ----------------------------------------------------------------------------
 """Tictactoe cog."""
-from typing import Final
+import abc
+import math
+from enum import Enum
+from io import BytesIO
+from typing import Final, NamedTuple
 
 import discord
 from discord import ButtonStyle, app_commands, ui
 from discord.ext import commands
 from discord.utils import utcnow
-from tictactoe import TicTacABC, TicTacEasy, TicTacHard
+from PIL import Image
 
 from main import CBot
 
@@ -48,6 +52,650 @@ ALLOWED_ROLES: Final = (
 CHANNEL_ID: Final = 969972085445238784
 
 MESSAGE: Final = "You must be at least level 1 to participate in the giveaways system and be in <#969972085445238784>."
+
+
+class PilOffsets(Enum):
+    """Enum for PIL Offsets."""
+
+    TOP_LEFT = (0, 0)
+    TOP_MIDDLE = (179, 0)
+    TOP_RIGHT = (355, 0)
+    MIDDLE_LEFT = (0, 179)
+    MIDDLE_MIDDLE = (179, 179)
+    MIDDLE_RIGHT = (355, 179)
+    BOTTOM_LEFT = (0, 357)
+    BOTTOM_MIDDLE = (179, 357)
+    BOTTOM_RIGHT = (355, 357)
+
+
+class GridPositions(Enum):
+    """Enum for Grid Positions."""
+
+    TOP_LEFT = "(0, 0)"
+    TOP_MIDDLE = "(0, 1)"
+    TOP_RIGHT = "(0, 2)"
+    MIDDLE_LEFT = "(1, 0)"
+    MIDDLE_MIDDLE = "(1, 1)"
+    MIDDLE_RIGHT = "(1, 2)"
+    BOTTOM_LEFT = "(2, 0)"
+    BOTTOM_MIDDLE = "(2, 1)"
+    BOTTOM_RIGHT = "(2, 2)"
+
+
+class Points(NamedTuple):
+    """NamedTuple for Points."""
+
+    participation: int
+    bonus: int
+
+
+class TicTacABC(abc.ABC):
+    """Abstract class for TicTac.
+
+    Methods
+    -------
+    display()
+        Render the board as a discord.File object.
+    move(x: int, y: int)
+        Parse a player's move.
+    check_win()
+        Check if the game has been won.
+    next()
+        Make an AI move.
+    """
+
+    @property
+    @abc.abstractmethod
+    def letter(self) -> str:
+        """Return the letter of the player."""
+
+    @property
+    @abc.abstractmethod
+    def points(self) -> Points:
+        """Return the points of the player."""
+
+    @abc.abstractmethod
+    def display(self) -> discord.File:
+        """Return an image of the board.
+
+        Returns
+        -------
+        discord.File
+            The image of the board, represented as a discord.File object.
+        """
+
+    @abc.abstractmethod
+    def move(self, x: int, y: int) -> bool:
+        """Record a move.
+
+        Parameters
+        ----------
+        x : int
+            The x position of the move.
+        y : int
+            The y position of the move.
+        """
+
+    @abc.abstractmethod
+    def check_win(self) -> int:
+        """Check if the player has won.
+
+        Returns
+        -------
+        int
+            0 if no one has won.
+            1 if the player has won.
+            -1 if the computer has won.
+        """
+
+    @abc.abstractmethod
+    def next(self) -> tuple[int, int]:
+        """Make a move, and return the cell that was played.
+
+        Returns
+        -------
+        tuple[int, int]
+            The x and y position of the move.
+        """
+
+
+class TicTacEasy(TicTacABC):
+    """Hard AI implementation of TicTacToe.
+
+    Parameters
+    ----------
+    pick : str
+        The letter of the player.
+    sz : int = 3
+        The size of the board.
+
+    Attributes
+    ----------
+    dim_sz : int
+        The size of the board.
+    board : list[list[str]]
+        The board.
+    pick : str
+        The letter of the player.
+
+    Methods
+    -------
+    display()
+        Render the board as a discord.File object.
+    move(x: int, y: int)
+        Parse a player's move.
+    check_win()
+        Check if the game has been won.
+    next()
+        Make an AI move.
+
+    See Also
+    --------
+    :class:`TicTacEasy`
+        An implementation with an easier AI.
+    :class:`TicTacABC`
+        The abstract class for TicTac. All other implementations inherit from this class, or a subclass of it.
+    """
+
+    def __init__(self, pick, sz=3) -> None:
+        self.pick = pick
+        self.dim_sz = sz
+        self.board = self.clear_board()
+
+    def clear_board(self) -> list[list[str]]:
+        """Clear the board.
+
+        Returns
+        -------
+        list[list[str]]
+            The cleared board.
+        """
+        board = [["blur" for _ in range(self.dim_sz)] for _ in range(self.dim_sz)]
+        # made a 3x3 by-default board
+        return board
+
+    @property
+    def letter(self) -> str:
+        """Return the letter of the player.
+
+        Returns
+        -------
+        str
+            The letter of the player.
+        """
+        return self.pick
+
+    @property
+    def points(self) -> Points:
+        """Return the points of the game.
+
+        Returns
+        -------
+        Points
+            The points of the game.
+        """
+        win_state = self.check_win()
+        if win_state == 1:
+            return Points(participation=1, bonus=1)
+        if win_state == 0:
+            return Points(participation=0, bonus=0)
+        return Points(participation=1, bonus=0)
+
+    # noinspection DuplicatedCode
+    def check_win(self) -> int:
+        """Check if a win state has been reached.
+
+        Returns
+        -------
+        int
+            -1 if no one has won.
+            1 if the player has won.
+            0 if the computer has won.
+        """
+        # 1 you won, 0 computer won, -1 tie
+        # Flag syntax -> first player no. ,
+        # User is Player#1 ;
+        # Check set 1 -> row and '\' diagonal & Check set 2 -> col and '/' diagonal
+
+        for i in range(0, self.dim_sz):  # Rows
+            flag11 = True
+            flag21 = True
+
+            flag12 = True
+            flag22 = True
+            for j in range(0, self.dim_sz):
+
+                ch2 = self.board[i][j]
+                ch1 = self.board[j][i]
+                # Row
+                if ch1 == self.pick:  # if it's mine, computer didn't make it
+                    flag21 = False
+                elif ch1 == "blur":  # if it's blank no one made it
+                    flag11 = False
+                    flag21 = False
+                else:
+                    flag11 = False  # else I didn't make it
+
+                if ch2 == self.pick:  # Same but for Col
+                    flag22 = False
+                elif ch2 == "blur":
+                    flag12 = False
+                    flag22 = False
+                else:
+                    flag12 = False
+
+            if flag11 is True or flag12 is True:  # I won
+                return 1
+            if flag21 is True or flag22 is True:  # Computer Won
+                return 0
+
+        # Diagonals#
+        flag11 = True
+        flag21 = True
+
+        flag12 = True
+        flag22 = True
+        for i in range(0, self.dim_sz):
+
+            ch2 = self.board[i][i]
+            ch1 = self.board[i][self.dim_sz - 1 - i]
+
+            if ch1 == self.pick:
+                flag21 = False
+            elif ch1 == "blur":
+                flag11 = False
+                flag21 = False
+            else:
+                flag11 = False
+
+            if ch2 == self.pick:
+                flag22 = False
+            elif ch2 == "blur":
+                flag12 = False
+                flag22 = False
+            else:
+                flag12 = False
+
+        if flag11 or flag12:
+            return 1
+        if flag21 or flag22:
+            return 0
+
+        return -1
+
+    def move(self, x: int, y: int) -> bool:
+        """Record a move.
+
+        Parameters
+        ----------
+        x : int
+            The x position of the move.
+        y : int
+            The y position of the move.
+
+        Returns
+        -------
+        bool
+            True if the move was successful, False otherwise.
+        """
+        move = self._move_record(x, y)
+        if isinstance(move, bool):
+            return move
+        return False
+
+    def display(self) -> discord.File:
+        """Return an image of the board.
+
+        Returns
+        -------
+        :class:`discord.File`
+            The image of the board, represented as a discord.File object.
+        """
+        grid = Image.open("media/tictactoe/grid.png", "r").convert("RGBA")
+        cross = Image.open("media/tictactoe/X.png", "r")
+        circle = Image.open("media/tictactoe/O.png", "r")
+        for i in range(0, self.dim_sz):
+            for j in range(0, self.dim_sz):
+                if self.board[i][j] == "X":
+                    grid.paste(cross, PilOffsets[GridPositions(f"({i}, {j})").name].value, cross)
+                elif self.board[i][j] == "O":
+                    grid.paste(circle, PilOffsets[GridPositions(f"({i}, {j})").name].value, circle)
+        _bytes = BytesIO()
+        grid.save(_bytes, format="PNG")
+        _bytes.seek(0)
+        return discord.File(_bytes, filename="tictactoe.png")
+
+    def _move_record(self, r, c) -> str | bool:
+        """Record a move.
+
+        Parameters
+        ----------
+        r : int
+            The x position of the move.
+        c : int
+            The y position of the move.
+
+        Returns
+        -------
+        str | bool
+            True on sucess, string error message on failure.
+        """
+        if r > self.dim_sz or c > self.dim_sz:
+            return "Out of Bounds"
+        if self.board[r][c] != "blur":
+            return "Spot Pre-Occupied"
+        self.board[r][c] = self.pick
+        return True
+
+    def next(self) -> tuple[int, int]:
+        """Make a move, and return the cell that was played.
+
+        Returns
+        -------
+        tuple[int, int]
+            The x and y position of the move.
+        """
+        move = self._next_move("O" if self.letter == "X" else "X")
+        if move["position"] is None:
+            return -1, -1
+        position = move["position"]
+        assert isinstance(position, tuple)  # skipcq: BAN-B101
+        self.board[position[0]][position[1]] = "O" if self.letter == "X" else "X"
+        return position
+
+    def _available_moves(self) -> list[tuple[int, int]]:
+        """Return a list of available moves.
+
+        Returns
+        -------
+        list[tuple[int, int]]
+            A list of available moves.
+        """
+        available_moves = []
+        for i in range(0, self.dim_sz):
+            for j in range(0, self.dim_sz):
+                if self.board[i][j] == "blur":
+                    available_moves.append((i, j))
+        return available_moves
+
+    def _empty_squares(self) -> int:
+        """Return the number of empty squares.
+
+        Returns
+        -------
+        int
+            The number of empty squares.
+        """
+        empty_squares = 0
+        for i in range(0, self.dim_sz):
+            for j in range(0, self.dim_sz):
+                if self.board[i][j] == "blur":
+                    empty_squares += 1
+        return empty_squares
+
+    # noinspection PyAssignmentToLoopOrWithParameter,DuplicatedCode
+    def _next_move(self, player: str) -> dict[str, tuple[int, int] | int | float | None]:
+        """Make a move, and return the cell that was played.
+
+        Returns
+        -------
+        tuple[int, int] | tuple[list[int], list[int]]
+            The cell that was played.
+        """
+        other_player = self.letter
+        max_player = "O" if self.letter == "X" else "X"  # yourself
+
+        # first we want to check if the previous move is a winner
+        if self.check_win() == 1:
+            return {
+                "position": None,
+                "score": 1 * (self._empty_squares() + 1)
+                if other_player == max_player
+                else -1 * (self._empty_squares() + 1),
+            }
+        if len(self._available_moves()) == 0:
+            return {"position": None, "score": 0}
+
+        if player == max_player:
+            best = {"position": None, "score": -math.inf}  # each score should maximize
+        else:
+            best = {"position": None, "score": math.inf}  # each score should minimize
+        for possible_move in self._available_moves():
+            self.board[possible_move[0]][possible_move[1]] = player
+            sim_score = self._next_move(other_player)  # simulate a game after making that move
+
+            # undo move
+            self.board[possible_move[0]][possible_move[1]] = "blur"
+            sim_score["position"] = possible_move  # this represents the move optimal next move
+
+            if player == max_player:  # X is max player
+                sim_score_val = sim_score["score"]
+                best_val = best["score"]
+                assert isinstance(sim_score_val, (float, int))  # skipcq: BAN-B101
+                assert isinstance(best_val, (float, int))  # skipcq: BAN-B101
+                if sim_score_val > best_val:
+                    best = sim_score
+            else:
+                sim_score_val = sim_score["score"]
+                best_val = best["score"]
+                assert isinstance(sim_score_val, (float, int))  # skipcq: BAN-B101
+                assert isinstance(best_val, (float, int))  # skipcq: BAN-B101
+                if sim_score_val < best_val:
+                    best = sim_score
+        return best
+
+
+class TicTacHard(TicTacEasy):
+    """Adaptation of TicTacHard to an easier version.
+
+    See Also
+    --------
+    :class:`TicTacHard`
+        For the full documentation and implementation.
+    """
+
+    @property
+    def points(self) -> Points:
+        """Return the points of the game.
+
+        Returns
+        -------
+        Points
+            The points of the game.
+        """
+        win_state = self.check_win()
+        if win_state == 1:
+            return Points(participation=2, bonus=3)
+        if win_state == 0:
+            return Points(participation=0, bonus=0)
+        return Points(participation=2, bonus=1)
+
+    def next(self) -> tuple[int, int]:
+        """Make a move, and return the cell that was played.
+
+        Returns
+        -------
+        tuple[int, int]
+            The x and y position of the move.
+        """
+        move_x, move_y = self._next_move_easy()
+        if isinstance(move_x, int) and isinstance(move_y, int):
+            return move_x, move_y
+        assert isinstance(move_x, list)  # skipcq: BAN-B101
+        assert isinstance(move_y, list)  # skipcq: BAN-B101
+        return move_x[0], move_y[0]
+
+    def _available_moves(self) -> list[tuple[int, int]]:
+        """Return a list of available moves.
+
+        Returns
+        -------
+        list[tuple[int, int]]
+            A list of available moves.
+        """
+        return [(i, j) for i in range(0, self.dim_sz) for j in range(0, self.dim_sz) if self.board[i][j] == "blur"]
+
+    def _win_spots(self, pick1: str, pick2: str) -> tuple[list[tuple[int, int]], list[tuple[int, int]]]:
+        """Return a list of winning moves.
+
+        Parameters
+        ----------
+        pick1 : str
+            The pick to check for the player.
+        pick2 : str
+            The pick to check for the computer.
+
+        Returns
+        -------
+        list[tuple[int, int]]
+            A list of winning moves.
+        """
+        win_spots1 = []
+        win_spots2 = []
+        for i in range(0, self.dim_sz):
+            for j in range(0, self.dim_sz):
+                if self.board[i][j] == "blur":
+                    self.board[i][j] = pick1
+                    if self.check_win() == 1:
+                        win_spots1.append((i, j))
+                    self.board[i][j] = pick2
+                    if self.check_win() == 0:
+                        win_spots2.append((i, j))
+                    self.board[i][j] = "blur"
+        return win_spots1, win_spots2
+
+    def _simple_move_checks(
+        self,
+        comp: list[tuple[int, int]],
+        player: list[tuple[int, int]],
+        available: list[tuple[int, int]],
+        comp_pick: str,
+    ) -> int | tuple[int, int]:
+        """Check if there is a simple move.
+
+        Parameters
+        ----------
+        comp : list[tuple[int, int]]
+            The computer's moves that would win the game for the computer.
+        player : list[tuple[int, int]]
+            The player's moves that would win the game for the player.
+        available : list[tuple[int, int]]
+            The available moves.
+        comp_pick : str
+            The pick for the computer.
+
+        Returns
+        -------
+        int | tuple[int, int]
+            The move to make, or 0 if there is no simple move.
+        """
+        if len(comp) != 0:
+            return comp[0]
+        if len(player) != 0:
+            self.board[player[0][0]][player[0][1]] = comp_pick
+            return player[0][0], player[0][1]
+        if len(available) == 1:
+            self.board[available[0][0]][available[0][1]] = comp_pick
+            return available[0][0], available[0][1]
+        if len(available) == 0:
+            return -1, -1
+        if self.board[0][0] == self.pick and self.board[2][2] == self.pick:
+            if self.board[0][2] == "blur":
+                self.board[0][2] = comp_pick
+                return 0, 2
+            if self.board[2][0] == "blur":
+                self.board[2][0] = comp_pick
+                return 2, 0
+        if self.board[0][2] == self.pick and self.board[2][0] == self.pick:
+            if self.board[0][0] == "blur":
+                self.board[0][0] = comp_pick
+                return 0, 0
+            if self.board[2][2] == "blur":
+                self.board[2][2] = comp_pick
+                return 2, 2
+        return 0
+
+    def _complex_move_checks(self, available: list[tuple[int, int]], comp_pick: str) -> int | tuple[int, int]:
+        """Check if there is a complex move.
+
+        Parameters
+        ----------
+        available : list[tuple[int, int]]
+            The available moves.
+        comp_pick : str
+            The pick for the computer.
+
+        Returns
+        -------
+        int | tuple[int, int]
+            The move to make, or 0 if there is no complex move.
+        """
+        c1, c2 = self.dim_sz // 2, self.dim_sz // 2
+        for i in range(c1 - 1, -1, -1):  # IN TO OUT
+            gap = c1 - i
+            # checking  - 4 possibilities at max
+            # EDGES
+            if (c1 - gap, c2 - gap) in available:
+                self.board[c1 - gap][c2 - gap] = comp_pick
+                return c1 - gap, c2 - gap
+            if (c1 - gap, c2 + gap) in available:
+                self.board[c1 - gap][c2 + gap] = comp_pick
+                return c1 - gap, c2 + gap
+            if (c1 + gap, c2 - gap) in available:
+                self.board[c1 + gap][c2 - gap] = comp_pick
+                return c1 + gap, c2 - gap
+            if (c1 + gap, c2 + gap) in available:
+                self.board[c1 + gap][c2 + gap] = comp_pick
+                return c1 + gap, c2 + gap
+
+            # Four Lines
+
+            for j in range(0, gap):
+                if (c1 - gap, c2 - gap + j) in available:  # TOP LEFT TO TOP RIGHT
+                    self.board[c1 - gap][c2 - gap + j] = comp_pick
+                    return c1 - gap, c2 - gap + j
+                if (
+                    c1 + gap,
+                    c2 - gap + j,
+                ) in available:  # BOTTOM LEFT TO BOTTOM RIGHT
+                    self.board[c1 + gap][c2 - gap + j] = comp_pick
+                    return c1 + gap, c2 - gap + j
+                if (c1 - gap, c2 - gap) in available:  # LEFT TOP TO LEFT BOTTOM
+                    self.board[c1 - gap + j][c2 - gap] = comp_pick
+                    return c1 - gap + j, c2 - gap
+                if (
+                    c1 - gap + j,
+                    c2 + gap,
+                ) in available:  # RIGHT TOP TO RIGHT BOTTOM
+                    self.board[c1 - gap + j][c2 + gap] = comp_pick
+                    return c1 - gap + j, c2 + gap
+        return 0
+
+    def _next_move_easy(self) -> tuple[int, int] | tuple[list[int], list[int]]:
+        """Make a move, and return the cell that was played.
+
+        Returns
+        -------
+        tuple[int, int] | tuple[list[int], list[int]]
+            The cell that was played.
+        """
+        available_moves = self._available_moves()  # will carry all available moves
+        comp_pick = "O"
+        if self.pick == "O":
+            comp_pick = "X"
+        # if player (user Wins), if computer (computer Wins)
+        player_win_spot, comp_win_spots = self._win_spots(self.pick, comp_pick)
+        simple_move = self._simple_move_checks(comp_win_spots, player_win_spot, available_moves, comp_pick)
+        if isinstance(simple_move, tuple):
+            return simple_move
+        c1, c2 = self.dim_sz // 2, self.dim_sz // 2
+        if (c1, c2) in available_moves:  # CENTER
+            self.board[c1][c2] = comp_pick
+            return c1, c2
+        complex_move = self._complex_move_checks(available_moves, comp_pick)
+        if isinstance(complex_move, tuple):
+            return complex_move
+        raise RuntimeError("No moves available")
 
 
 class TicTacView(ui.View):
